Add files via upload
This project is based on Web Development And its Applications. The main objective of this project is to learn the implementation of HTML, CSS and JavaScript.The basic webpage of this project is created using HTML and styling of the webpage is done using CSS and database is done using Py and Django.